Added doubly Linked Lists, modified singly linked list
Changed the singly linked list so that the head and tail would point to
the same node for lists with only one item. This was done because doubly
linked list need this to be true to print a list backwards, and I want
singly and doubly linked list to be consistent with each other.